v2/stt-service/requirements-simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
pydantic==2.5.0
numpy==1.24.3
soundfile==0.12.1
prometheus-client==0.19.0
structlog==23.2.0
openai-whisper==20231117 